Create Callback,PromisemAsync and Await in ES6
What is callback in JavaScript?

Ans = A callback is a function passed as an argument to another function. This technique allows a function to call another function. A callback function can run after another function has finished.

What is Promise in JavaScript?

Ans =  Image result for what is promise in es6
The Promise object represents the eventual completion (or failure) of an asynchronous operation and its resulting value. To learn about the way promises work and how you can use them, we advise you to read Using promises first

Q. What are async and await in Javascript ES6?

Async and Await both are considered as special keywords which are provided by ES6 in order to perform some asynchronous data operations. Even synchronous operations could also be performed using these keywords